Wire in the select boxes
On a select change we now recall the code to generate routes from csv
data and redraw the map.  Next step is to use the filters when
generating the routes.

Also put code in to distinguish between single and multiselects (but we
still need to make the UI support multiselect)