Refactor csv load error reporting
We now have a structure to allow for future error messages (beyond
unrecognised country codes) and it is encapsulated in trademapper.csv -
it stores the errors, and loadErrorsToStrings() will return an array of
strings describing any errors.

trademapper will then display those errors.

Also use trim() in a few places to increase robustness.