finish disable/enable year range functionality
when the year slider is enabled, disable the year range form element,
and put some text to make it clear that is disabled.

Also as soon as the switch is changed, change the data displayed from
data for all years, to data for just one year.

And vice versa for both.