Apply zoom behaviour to whole SVG. Applying to zoomgroup was causing jittery panning
The 'zoom' event is fired many times when the user is panning and zooming. If zoom behaviour is applied to the same element as the one we are transforming their are two conflicting sources of updates to the d3.event.translate values (one from the .attr('transform',...) and one from the zoom behaviour).