Basic image export to PNG working
To make this work, I moved styling out of the stylesheet into
the SVG itself. If this is not done, the exported SVG loses
styling information and doesn't render correctly.

The same needs to be done for trade routes: at the moment, only
the base map renders correctly.

I also simplified the disputed territories, colouring them a
pale transparent red rather than using hatching. This reduces
issues which may be introduced by attempting to use background
images and exporting those into the PNG.