CSV fetch fails due to API change
This wasn't tested for when upgrading to d3 v5, as tests were
confined to uploading a CSV file.

The d3 API changed in v5, so that d3.xhr() no longer exists.
Its replacement is d3.text(), so use that.

The API has also become Promises-based, so we can clean
up some of the code which was waiting for responses using
timeouts, too.

Note that d3.csv() also exists now, which could potentially make
some of the code in trademapper.csv.js redundant.