Fix SVG render destroying data bindings
SVG for the overlay was being rendered to a string then appended
to the existing HTML (as a string) inside the SVG.

This was causing the binding between the SVG and the data to
be destroyed, as it lost any bindings between the DOM elements
and their underlying data source (established through d3).

Return to rendering the overlay using d3 data binding, rather than
a string, to retain the binding from the map to the data.

This means that the country colours are retained, and that the
play button works.