Report errors when loading topojson
We don't have a generic way to show errors which is application-wide,
so we can't really show errors caused by loading an overlay from
the query string.

When we have some UX for this, we can show errors in the appropriate
tab (where the overlay file is specified), as is done for CSV errors.