Pause playback while exporting gif
Because the gif export uses the SVG element from the DOM
directly, we need to ensure that it's not changing while the
export is going on.

Explicitly pause the playback of the map slider before starting
the export, then restore the play state after the export.

This required adding play/pause commands to the yearslider, which
isn't a bad thing anyway.