Styling fix for overlong tabs
Working on the port import highlighted an issue where field names
are long in the import CSV: the filter tab gets too long for the
browser window, which causes the whole page to scroll; in turn,
this causes the map to scroll, which messes with the scrollwheel
to zoom functionality.

By setting a max tab height and adding overflow-y, we can just
scroll the tab content when necessary, which means we don't
break scroll on the map.

On older browsers which don't understand vh units for height,
the bad behaviour will remain. However, I was unable to find a
combination of CSS styles which worked that didn't use vh units.
So at least nothing is worse than before.