Allow multiple location columns for same order
Code previously prevented multiple columns being used for the
same trade node type, e.g. it wasn't possible to have
a country code, port code and lat/long specified as the origin
in a single row of a dataset.

Remove this constraint so that multiple columns can be assigned
to the same location order, and choose the most-specific column
value when creating routes.

The location order now also acts as a grouping for location
columns: where multiple location columns have the same order
(group), the most specific point is used to construct routes.

This required rewriting a lot of the tests. The position of
a location in the ordering is now encoded into the point, so
some of the tests which added the same points but arranged them
differently were failing. Additional points were added with
correct ordering to make these tests pass again.