Sort out mess caused by route role combining
The implementation for colouring countries ended up with roles
being incorrectly combined along routes.

Revert the code to what it was previously and add a separate
simpler method to get all the country codes along a route (both
for countries and ports). Then use this new method to do the
country colouring.