Go back to gif.js library; fix GIF export
This library performs far better than Animated_GIF.

The colour mapping isn't quite as good, but it can handle
the large gif exports we're now doing, which generally give
a higher fidelity movie. It is also faster.

Also fix some styling on the SVG so the exported GIF doesn't
have a black border, but instead has a background filled with
the ocean colour.